Refactor: LottoGameService 클래스 service 패키지로 이동
